DRY client out again with common spawn thread
