+++ nameko_grpc/client.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-import sys
 import threading
 import time
 from logging import getLogger
@@ -14,6 +13,7 @@ from nameko_grpc.errors import GrpcError
 from nameko_grpc.inspection import Inspector
 from nameko_grpc.ssl import SslConfig
 from nameko_grpc.timeout import bucket_timeout
+from nameko_grpc.threading import target_with_callback
 
 
 log = getLogger(__name__)
@@ -187,19 +187,5 @@ class Client(ClientBase):
         return Proxy(self)
 
     def spawn_thread(self, target, args=(), kwargs=None, name=None, callback=None):
-
-        if kwargs is None:
-            kwargs = {}
-
-        def execute():
-            try:
-                res = target(*args, **kwargs)
-            except Exception:
-                res = None
-                exc_info = sys.exc_info()
-            else:
-                exc_info = None
-            if callback:
-                callback(res, exc_info)
-
+        execute = target_with_callback(target, args, kwargs, callback)
         threading.Thread(target=execute, name=name).start()
